move classes for debug to their test files
